@@ -1,23 +1,22 @@
-import React from 'react';
-import {
-  makeStyles,
-  createStyles,
-  AppBar,
-  Toolbar,
-  IconButton
-} from '@material-ui/core';
-import { Undo, Save, Facebook, Twitter } from '@material-ui/icons';
+import AppBar from "@mui/material/AppBar";
+import Toolbar from "@mui/material/Toolbar";
+import Undo from "@mui/icons-material/Undo";
+import Save from "@mui/icons-material/Save";
+import Facebook from "@mui/icons-material/Facebook";
+import Twitter from "@mui/icons-material/Twitter";
+import { makeStyles, createStyles } from "@mui/styles";
+import IconButton from "@mui/material/IconButton";
 
 const useStyles = makeStyles(() =>
   createStyles({
     // offset: theme.mixins.toolbar,
     appBar: {
-      top: 'auto',
-      bottom: 0
+      top: "auto",
+      bottom: 0,
     },
     toolbar: {
-      alignSelf: 'center'
-    }
+      alignSelf: "center",
+    },
   })
 );
 
@@ -25,23 +24,21 @@ export default function Footer() {
   const classes = useStyles();
 
   return (
-    <>
-      <AppBar position="sticky" color="primary" className={classes.appBar}>
-        <Toolbar className={classes.toolbar}>
-          <IconButton color="inherit" aria-label="">
-            <Undo />
-          </IconButton>
-          <IconButton color="inherit" aria-label="">
-            <Save />
-          </IconButton>
-          <IconButton color="inherit" aria-label="">
-            <Facebook />
-          </IconButton>
-          <IconButton color="inherit" aria-label="">
-            <Twitter />
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-    </>
+    <AppBar position="sticky" color="primary" className={classes.appBar}>
+      <Toolbar className={classes.toolbar}>
+        <IconButton color="inherit" aria-label="">
+          <Undo />
+        </IconButton>
+        <IconButton color="inherit" aria-label="">
+          <Save />
+        </IconButton>
+        <IconButton color="inherit" aria-label="">
+          <Facebook />
+        </IconButton>
+        <IconButton color="inherit" aria-label="">
+          <Twitter />
+        </IconButton>
+      </Toolbar>
+    </AppBar>
   );
 }
