@@ -1,16 +1,15 @@
-import React from 'react';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
-import { makeStyles } from '@material-ui/core/styles';
-import MoodIcon from '@material-ui/icons/Mood';
+import AppBar from "@mui/material/AppBar";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import IconButton from "@mui/material/IconButton";
+import { makeStyles } from "@mui/styles";
+import MoodIcon from "@mui/icons-material/Mood";
 
 const useStyles = makeStyles(() => ({
   typography: {
     flexGrow: 1,
-    alignSelf: 'center'
-  }
+    alignSelf: "center",
+  },
 }));
 
 export default function Header() {
