@@ -1,4 +1,4 @@
-interface FixedLengthArray<T extends any, L extends number> extends Array<T> {
+interface FixedLengthArray<T, L extends number> extends Array<T> {
   0: T;
   length: L;
 }
